# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Vietnam Smart Golf Platform API
  description: |
    REST API for the Vietnam Smart Golf Platform.
    Covers identity, course catalog, geospatial, round/score, weather, correction workflow, and audit.
  version: 0.1.0
  contact:
    name: VSP Platform Team
  license:
    name: Proprietary
servers:
  - url: https://api.vsp.local
    description: Local development
  - url: https://api.vsp-staging.example.com
    description: Staging environment
  - url: https://api.vsp.example.com
    description: Production environment
paths:
...
  /courses/{courseId}/packages/build:
    post:
      operationId: triggerPackageBuild
      summary: Trigger a package build for a course
      description: |
        Creates and queues a PackageBuildJob for the given course and data version.
        Idempotent: if a non-terminal job already exists for this course, returns
        that job ID without creating a duplicate (HTTP 200, not 201).
        Requires COURSE_ADMIN or higher role.
      tags: [Courses]
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Course ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerBuildRequest'
      responses:
        '200':
          description: Build already in progress — returns existing job ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerBuildResponse'
        '201':
          description: Build job created and queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerBuildResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden — COURSE_ADMIN or higher required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Course or data version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Course not published — cannot build package
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            example:
              code: COURSE_NOT_PUBLISHED
              message: Cannot build a package for an unpublished course
              correlationId: "550e8400-e29b-41d4-a716-446655440000"
  /courses/{courseId}/packages/build/jobs:
    get:
      operationId: listPackageBuildJobs
      summary: List build jobs for a course
      description: |
        Returns all build jobs for a course, newest first.
        Used by the portal to display job history and polling.
        Requires COURSE_ADMIN or higher role.
      tags: [Courses]
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Course ID
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page number (0-indexed)
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          description: Page size
      responses:
        '200':
          description: Paginated list of build jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageBuildJobListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden — COURSE_ADMIN or higher required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /courses/{courseId}/packages/build/jobs/{jobId}:
    get:
      operationId: getPackageBuildJob
      summary: Get a specific build job
      description: |
        Returns details for a specific build job.
        Used by the portal for polling and retry.
        Requires COURSE_ADMIN or higher role.
      tags: [Courses]
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Course ID
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Build job ID
      responses:
        '200':
          description: Build job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageBuildJobDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden — COURSE_ADMIN or higher required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Build job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            example:
              code: BUILD_NOT_FOUND
              message: Build job not found
              correlationId: "550e8400-e29b-41d4-a716-446655440000"

  # ========================
  # Packages endpoints
  # ========================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    Error:
      $ref: './schemas/common.yaml#/Error'
...
    PackageBuildJobDto:
      $ref: './schemas/course.yaml#/PackageBuildJobDto'
    TriggerBuildRequest:
      $ref: './schemas/course.yaml#/TriggerBuildRequest'
    TriggerBuildResponse:
      $ref: './schemas/course.yaml#/TriggerBuildResponse'
    PackageBuildJobListResponse:
      $ref: './schemas/course.yaml#/PackageBuildJobListResponse'

    # Course version publish schemas (Story 8.3 PUBLISH-CONTRACT)
...